Added microbit_square.py
Microbit_square.py is the code that runs on the microbit, because square uses imports that the microbit doesn't support. This presents the images in a slideshow format. Press A to move to the previous slide, and press b to move to the next slide. (Little buggy, but I'll fix that later)